refactor(Database): ♻️ Move database folder into structures
